Refactored code and changed name of UnitTests to UnitTests_Controllers
